requirements.txt
boto3==1.34.104
botocore==1.34.104
cli_helpers==2.3.1
click==8.1.7
configobj==5.0.8
confluent-kafka==2.4.0
jmespath==1.0.1
logging==0.4.9.6
numpy==1.26.4
pgcli==4.1.0
pgspecial==2.1.1
prompt-toolkit==3.0.43
psycopg==3.1.19
psycopg2-binary==2.9.9
Pygments==2.18.0
python-dateutil==2.9.0.post0
s3transfer==0.10.1
setproctitle==1.3.3
six==1.16.0
sqlparse==0.5.0
tabulate==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13